+++ src/v/security/mtls.cc
@@ -27,11 +27,6 @@ parse_rules(std::optional<std::vector<ss::sstring>> unparsed_rules);
 
 } // namespace detail
 
-std::ostream& operator<<(std::ostream& os, const rule& r) {
-    fmt::print(os, "{}", r);
-    return os;
-}
-
 std::ostream& operator<<(std::ostream& os, const principal_mapper& p) {
     fmt::print(os, "{}", p);
     return os;
@@ -57,30 +52,6 @@ std::optional<ss::sstring> principal_mapper::apply(std::string_view sv) const {
 
 // explicit instantiations so as to avoid bringing in <fmt/ranges.h> in the
 // header, whch breaks compilation in another part of the codebase.
-template<>
-typename fmt::basic_format_context<fmt::appender, char>::iterator
-fmt::formatter<security::tls::rule, char, void>::format<
-  fmt::basic_format_context<fmt::appender, char>>(
-  security::tls::rule const& r,
-  fmt::basic_format_context<fmt::appender, char>& ctx) const {
-    if (r._is_default) {
-        return fmt::format_to(ctx.out(), "DEFAULT");
-    }
-    fmt::format_to(ctx.out(), "RULE:");
-    if (r._pattern.has_value()) {
-        fmt::format_to(ctx.out(), "{}", *r._pattern);
-    }
-    if (r._replacement.has_value()) {
-        fmt::format_to(ctx.out(), "/{}", *r._replacement);
-    }
-    if (r._to_lower) {
-        fmt::format_to(ctx.out(), "/L");
-    } else if (r._to_upper) {
-        fmt::format_to(ctx.out(), "/U");
-    }
-    return ctx.out();
-}
-
 template<>
 typename fmt::basic_format_context<fmt::appender, char>::iterator
 fmt::formatter<security::tls::principal_mapper, char, void>::format<
